query-service/requirements.txt
fastapi
uvicorn[standard]
pinecone
openai
sentence-transformers
pydantic-settings 